Updated ZIP library
Added some debugging to ePub decryption code. Apparently, some files appeared to be corrupted, so they decompressing to zero-length files, which causes some nasty behavior of Adobe-based readers (ADE and all devices that use it's SDK) all the way up to BSoD and reboot of devices :(

Funny thing is that those files WITH DRM in place are being read without issues.